Add team selection for notebooks
