+++ indigo-frontend/src/app/pages/notebook/notebook-info/notebook-info.component.ts
@@ -4,15 +4,21 @@ import { NotebookDetail } from '@/core/types/entities/notebook-detail.i';
 import { CommonModule } from '@angular/common';
 import { Component, inject } from '@angular/core';
 import { NotebookService } from '../../../../core/services/notebook/notebook.service';
+import { TeamComponent, TeamComponentConfig } from '@/core/components/common/team/team.component';
 
 @Component({
     selector: 'eln-notebook-info',
     standalone: true,
-    imports: [CommonModule, ButtonComponent, CardComponent],
+    imports: [CommonModule, ButtonComponent, CardComponent, TeamComponent],
     templateUrl: './notebook-info.component.html',
 })
 export class NotebookInfoComponent {
     private store = inject(NotebookService);
+
+    notebookTeamConfig: TeamComponentConfig = {
+      title: 'Notebook Team',
+      buildAccessEndpoint: (id: string) => `notebooks/${id}/access`,
+    };
 
     get notebook(): NotebookDetail | null {
         return this.store.notebook();
