+++ indigo-frontend/src/app/pages/project/project-info/project-info.component.html
@@ -1,69 +1,61 @@
 @if (isLoading) {
-  <div class="flex items-center justify-center py-8">
-    <span>Loading project...</span>
-  </div>
+<div class="flex items-center justify-center py-8">
+  <span>Loading project...</span>
+</div>
 } @else if (hasError) {
-  <div class="flex items-center justify-center py-8">
-    <div class="text-center">
-      <span>There was an error loading the project, please try again later</span>
-    </div>
+<div class="flex items-center justify-center py-8">
+  <div class="text-center">
+    <span>There was an error loading the project, please try again later</span>
   </div>
+</div>
 } @else if (project) {
-  <div class="grid grid-cols-[1fr_auto] gap-4 pt-4">
-    <eln-card>
-      <div class="flex items-center justify-between pb-3 border-b border-b-neutral-200">
-        <h2 class="font-semibold text-base">Project Description</h2>
-        <eln-button variant="blue-outline" classList="text-primary-400" (click)="openEditDialog()">
-          <em class="indicon-edit"></em>
-          Edit
-        </eln-button>
+<div class="grid grid-cols-[1fr_auto] gap-4 pt-4">
+  <eln-card>
+    <div class="flex items-center justify-between pb-3 border-b border-b-neutral-200">
+      <h2 class="font-semibold text-base">Project Description</h2>
+      <eln-button variant="blue-outline" classList="text-primary-400" (click)="openEditDialog()">
+        <em class="indicon-edit"></em>
+        Edit
+      </eln-button>
+    </div>
+
+    <div class="space-y-6 mt-3 text-sm">
+      <div>
+        <h3 class="font-semibold text-neutral-800 mb-2">Project Name</h3>
+        <p class="text-neutral-1000">{{ project.name }}</p>
       </div>
 
-      <div class="space-y-6 mt-3 text-sm">
-        <div>
-          <h3 class="font-semibold text-neutral-800 mb-2">Project Name</h3>
-          <p class="text-neutral-1000">{{ project.name }}</p>
+      <div>
+        <h3 class="text-neutral-800 mb-2 font-semibold">Project Keywords</h3>
+        <div class="flex flex-wrap gap-2">
+          @for (keyword of project.keywords; track keyword) {
+          <eln-chip>{{ keyword }}</eln-chip>
+          }
         </div>
-
-        <div>
-          <h3 class="text-neutral-800 mb-2 font-semibold">Project Keywords</h3>
-          <div class="flex flex-wrap gap-2">
-            @for (keyword of project.keywords; track keyword) {
-              <eln-chip>{{ keyword }}</eln-chip>
-            }
-          </div>
-        </div>
-
-        <!-- TODO Show literature references (empty by the moment) -->
-        <!-- <div>
-          <h3 class="text-neutral-800 mb-2 font-semibold">Literature</h3>
-          <p class="text-neutral-400">-</p>
-        </div> -->
-
-        @if (project.description || project.details) {
-          <div>
-            <h3 class="text-neutral-800 mb-2 font-semibold">Project Details</h3>
-            <p class="text-neutral-1000" [innerHTML]="project.description"></p>
-            <p class="text-neutral-1000 mt-4">{{ project.details }}</p>
-          </div>
-        }
-        @if (project.attachments.length > 0) {
-          <div>
-            <h3 class="text-neutral-800 mb-2 font-semibold">Attachments</h3>
-            <div class="space-y-2 grid">
-              @for (file of project.attachments; track file) {
-                <eln-attachment
-                  [attachment]="file"
-                  [projectId]="project.id"
-                  (attachmentDeleted)="onAttachmentDeleted($event)"
-                />
-              }
-            </div>
-          </div>
-        }
-        <eln-file-upload (filesSelected)="onUpload($event)" [uploadingFile]="isUploadingAttachment" [withPreview]="false"/>
       </div>
-    </eln-card>
-    <eln-team [team]="project.acl" [project]="project"/>
-  </div>
+
+      @if (project.description || project.details) {
+      <div>
+        <h3 class="text-neutral-800 mb-2 font-semibold">Project Details</h3>
+        <p class="text-neutral-1000" [innerHTML]="project.description"></p>
+        <p class="text-neutral-1000 mt-4">{{ project.details }}</p>
+      </div>
+      }
+      @if (project.attachments.length > 0) {
+      <div>
+        <h3 class="text-neutral-800 mb-2 font-semibold">Attachments</h3>
+        <div class="space-y-2 grid">
+          @for (file of project.attachments; track file) {
+          <eln-attachment [attachment]="file" [projectId]="project.id"
+            (attachmentDeleted)="onAttachmentDeleted($event)" />
+          }
+        </div>
+      </div>
+      }
+      <eln-file-upload (filesSelected)="onUpload($event)" [uploadingFile]="isUploadingAttachment"
+        [withPreview]="false" />
+    </div>
+  </eln-card>
+  <eln-team [team]="project.acl" [entityId]="project.id" [config]="projectTeamConfig" />
+</div>
 }
