+++ indigo-frontend/src/core/components/common/team/team.component.html
@@ -1,19 +1,14 @@
 <eln-card classList="h-full">
   <div class="flex items-center justify-between border-b border-b-neutral-200 pb-3">
     <div class="flex items-center gap-2">
-      <h2 class="font-semibold text-base">Team</h2>
+      <h2 class="font-semibold text-base">{{ title }}</h2>
       <eln-counter [count]="team.length" variant="grey" size="large" />
     </div>
-    <!-- TODO enable when sharing feature is implemented -->
-    <!-- <eln-button variant="blue-outline" classList="text-primary-400" [disabled]="true">
-      <em class="indicon-share"></em>
-      Share
-    </eln-button> -->
   </div>
 
   <div class="flex items-center justify-between py-3 gap-2">
     <div class="flex-1">
-  <ng-select [items]="userSuggestions" bindLabel="displayName" bindValue="id" [multiple]="true"
+      <ng-select [items]="userSuggestions" bindLabel="displayName" bindValue="id" [multiple]="true"
         [closeOnSelect]="false" [searchable]="true" placeholder="Add users..." [(ngModel)]="selectedUserIds"
         (change)="onUsersSelectedIds(selectedUserIds)" [clearable]="true" class="w-[300px]">
         <ng-template ng-option-tmp let-item="item" let-index="index">
@@ -39,22 +34,19 @@
 
   <div class="space-y-2 py-3">
     @for (member of team; track member) {
-    <div
-      class="grid grid-cols-[auto_1fr_auto_auto] gap-4 items-center p-2 hover:bg-neutral-100 rounded-lg text-neutral-1000">
-      <!-- TODO add avatar to ACL in project response -->
-      <!-- <eln-avatar [img]="member.avatar" classList="w-8 h-8" /> -->
-      <p class="text-neutral-1000">{{ member.displayName }}</p>
-      <p class="text-neutral-800 overflow-hidden text-ellipsis whitespace-nowrap max-w-[148px] text-ellipsis">
-        {{ member.username }}
-      </p>
-      <div class="flex items-center gap-2">
-        <eln-copy [text]="member.username" />
-        <eln-dropdown-menu class="[&_.dd-trigger]:text-primary-400 [&_.dd-container]:min-w-34" [controlled]="true"
-          [selected]="member.level | normalizeLabel" [items]="aclLevelOptions"
-          (itemSelected)="updateAclLevel(member, $event)"
-          [disabled]="isInmutableLevel(member.level) || isAclMemberLoading(member)" />
+      <div class="grid grid-cols-[auto_1fr_auto_auto] gap-4 items-center p-2 hover:bg-neutral-100 rounded-lg text-neutral-1000">
+        <p class="text-neutral-1000">{{ member.displayName }}</p>
+        <p class="text-neutral-800 overflow-hidden text-ellipsis whitespace-nowrap max-w-[148px] text-ellipsis">
+          {{ member.username }}
+        </p>
+        <div class="flex items-center gap-2">
+          <eln-copy [text]="member.username" />
+          <eln-dropdown-menu class="[&_.dd-trigger]:text-primary-400 [&_.dd-container]:min-w-34" [controlled]="true"
+            [selected]="member.level | normalizeLabel" [items]="aclLevelOptions"
+            (itemSelected)="updateAclLevel(member, $event)"
+            [disabled]="isInmutableLevel(member.level) || isAclMemberLoading(member) || config.disableLevelChange?.(member)" />
+        </div>
       </div>
-    </div>
     }
   </div>
 </eln-card>
